refactor: extract shared useSignupForm hook from duplicated signup logic
- New hooks/useSignupForm.ts with all signup logic (validation, API calls,
  GTM, identify/group, submission relay, error handling)
- TeamsVariant.tsx now consumes useSignupForm({ source: 'teams' })
- SignupModal.tsx now consumes useSignupForm({ source: 'homepage-modal' })
- Single source of truth for signup flow across both paths
